Make SerialCDC inherit from HardwareSerial
As suggested by @thiagoralves

+++ cores/arduino/USB/CDC.cpp
@@ -201,7 +201,7 @@ void Serial_::begin(uint32_t /* baud_count */)
 	// uart config is ignored in USB-CDC
 }
 
-void Serial_::begin(uint32_t /* baud_count */, uint8_t /* config */)
+void Serial_::begin(uint32_t /* baud_count */, uint16_t /* config */)
 {
 	// uart config is ignored in USB-CDC
 }
